refactor: DiaryRepository
