+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/diary/exception/DiaryExceptionTest.java
@@ -1,6 +1,6 @@
 package com.startingblue.fourtooncookie.diary.exception;
 
-import com.startingblue.fourtooncookie.diary.domain.DiaryRepository;
+import com.startingblue.fourtooncookie.diary.DiaryRepository;
 import com.startingblue.fourtooncookie.diary.service.DiaryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
